backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.14
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
httpx==0.26.0
python-dotenv==1.0.0
asyncpg==0.29.0
aiosqlite==0.20.0
